[1단계] chore : 패키지 생성 및 명명 변경, 의존성 추가
1. domain/reservation 패키지 생성
2. 콘솔앱 이름 변경
3. 롬복 의존성 추가
