+++ build.gradle
@@ -1,5 +1,6 @@
 plugins {
     id 'java'
+    id "io.freefair.lombok" version "6.5.1"
 }
 
 group 'nextstep'
