+++ src/main/java/nextstep/domain/reservation/Reservation.java
@@ -1,0 +1,34 @@
+package nextstep.domain.reservation;
+
+import java.time.LocalDate;
+import java.time.LocalTime;
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.RequiredArgsConstructor;
+import lombok.ToString;
+import lombok.experimental.FieldDefaults;
+
+@RequiredArgsConstructor
+@EqualsAndHashCode
+@ToString
+@Builder(toBuilder = true)
+@FieldDefaults(makeFinal = true, level = AccessLevel.PRIVATE)
+public class Reservation {
+
+  LocalDate date;
+  LocalTime time;
+  String name;
+
+  public LocalDate getDate() {
+    return date;
+  }
+
+  public LocalTime getTime() {
+    return time;
+  }
+
+  public String getName() {
+    return name;
+  }
+}
